requirements.txt
paho-mqtt
pigpio
RPi.GPIO